various improvments
- create schema only on one worker
- add typeorm verbose falg "DbDebug"
- make all phrases in search mandatory
