+++ quix-frontend/Dockerfile
@@ -6,7 +6,8 @@ RUN npm install
 WORKDIR /shared
 RUN npm install
 WORKDIR /service
-RUN npm install && npm run build && npm install -g pm2 && (rm -f .env || true)
+RUN npm install
+RUN npm run build && npm install -g pm2 && (rm -f .env || true)
 
 EXPOSE 3000
 ENTRYPOINT ["pm2-runtime", "start", "ecosystem.config.js"]
